Correction des Bug sur slider

+++ src/features/sliders/sliders.service.ts
@@ -14,55 +14,55 @@ export class SlidersService {
   constructor(
     @InjectModel(Slider.name) private readonly sliderModel: Model<Slider>,
   ) { }
-async create(
-  createSliderDto: CreateSliderDto,
-  files: { newImage?: Express.Multer.File[] }
-) {
-  const file = files.newImage?.[0];
-  let fullPath = '';
-  let relativePath = '';
+  async create(
+    createSliderDto: CreateSliderDto,
+    files: { newImage?: Express.Multer.File[] }
+  ) {
+    const file = files.newImage?.[0];
+    let fullPath = '';
+    let relativePath = '';
 
-  try {
-    // 1. Vérifications initiales
-    const existingSlider = await this.sliderModel.findOne({ title: createSliderDto.title });
-    if (existingSlider) return ApiResponse.error('Le slider existe déjà');
+    try {
+      // 1. Vérifications initiales
+      const existingSlider = await this.sliderModel.findOne({ title: createSliderDto.title });
+      if (existingSlider) return ApiResponse.error('Le slider existe déjà');
 
-    const existingOrder = await this.sliderModel.findOne({ order: createSliderDto.order });
-    if (existingOrder) return ApiResponse.error('Cet ordre est déjà utilisé');
+      const existingOrder = await this.sliderModel.findOne({ order: createSliderDto.order });
+      if (existingOrder) return ApiResponse.error('Cet ordre est déjà utilisé');
 
-    // 2. Gestion du fichier image
-    if (file) {
-      const uploadDir = './storage/sliders';
-      if (!fs.existsSync(uploadDir)) {
-        fs.mkdirSync(uploadDir, { recursive: true });
+      // 2. Gestion du fichier image
+      if (file) {
+        const uploadDir = './storage/sliders';
+        if (!fs.existsSync(uploadDir)) {
+          fs.mkdirSync(uploadDir, { recursive: true });
+        }
+        const uniqueSuffix = `${Date.now()}-${Math.round(Math.random() * 1e9)}`;
+        const fileName = `slider-${uniqueSuffix}${path.extname(file.originalname)}`;
+
+        fullPath = path.join(uploadDir, fileName);
+        relativePath = `storage/sliders/${fileName}`;
+
+        fs.writeFileSync(fullPath, file.buffer);
+      };
+      console.log(createSliderDto);
+
+      // 3. Création de l'instance et sauvegarde en BDD
+      const newSlider = new this.sliderModel({
+        ...createSliderDto,
+        image: relativePath, // On stocke le chemin relatif
+      });
+      console.log(newSlider);
+      const savedSlider = await newSlider.save();
+
+      return ApiResponse.success('Slider créé avec succès', savedSlider);
+
+    } catch (error) {
+      if (fullPath && fs.existsSync(fullPath)) {
+        fs.unlinkSync(fullPath);
       }
-
-      const uniqueSuffix = `${Date.now()}-${Math.round(Math.random() * 1e9)}`;
-      const fileName = `slider-${uniqueSuffix}${path.extname(file.originalname)}`;
-      
-      fullPath = path.join(uploadDir, fileName);
-      relativePath = `storage/sliders/${fileName}`;
-
-      fs.writeFileSync(fullPath, file.buffer);
+      return ApiResponse.error('Erreur lors de la création du slider');
     }
-
-    // 3. Création de l'instance et sauvegarde en BDD
-    const newSlider = new this.sliderModel({
-      ...createSliderDto,
-      image: relativePath, // On stocke le chemin relatif
-    });
-
-    const savedSlider = await newSlider.save();
-
-    return ApiResponse.success('Slider créé avec succès', savedSlider);
-
-  } catch (error) {
-    if (fullPath && fs.existsSync(fullPath)) {
-      fs.unlinkSync(fullPath);
-    }
-    return ApiResponse.error('Erreur lors de la création du slider');
   }
-}
   //Récuperation des sliders sans order 1
   async findAll() {
     try {
@@ -88,7 +88,7 @@ async create(
     }
   }
   async update(
-    slug: string,
+    idSlider: string,
     updateSliderDto: UpdateSliderDto,
     files: { newImage?: Express.Multer.File[] },
   ) {
@@ -98,7 +98,7 @@ async create(
 
     try {
       // 1. Trouver le slider actuel
-      const slider = await this.sliderModel.findOne({ slug });
+      const slider = await this.sliderModel.findById(idSlider);;
       if (!slider) {
         return ApiResponse.error('Slider introuvable');
       }
@@ -122,21 +122,25 @@ async create(
         // Mise à jour du chemin dans le DTO
         updateSliderDto.newImage = newRelativePath;
       }
-      const updatedSlider = await this.sliderModel.findOneAndUpdate(
-        { slug },
+      const updatedSlider = await this.sliderModel.findByIdAndUpdate(
+        idSlider, // Mongoose sait qu'il s'agit de l'ID technique
         { $set: updateSliderDto },
-        { new: true },
+        { new: true }, // Pour récupérer l'objet APRÈS modification
       );
+
+      if (!updatedSlider) {
+        return ApiResponse.error('Slider introuvable, mise à jour impossible.');
+      }
+
+      // La suite de votre logique pour l'image...
       if (file && oldImagePath) {
         const oldFullRootPath = path.join(process.cwd(), oldImagePath);
         if (fs.existsSync(oldFullRootPath)) {
           fs.unlinkSync(oldFullRootPath);
         }
       }
-      return ApiResponse.success(
-        'Slider mis à jour avec succès',
-        updatedSlider
-      );
+
+      return ApiResponse.success('Slider mis à jour avec succès', updatedSlider);
 
     } catch (error) {
       if (newRelativePath) {
@@ -146,30 +150,26 @@ async create(
       return ApiResponse.error('Erreur lors de la mise à jour');
     }
   }
-  async remove(slug: string) {
+  async remove(idSlider: string) {
     try {
-      const slider = await this.sliderModel.findOne({ slug });
+      // 1. Récupérer le slider pour avoir le chemin de l'image
+      const slider = await this.sliderModel.findById(idSlider);
+
       if (!slider) {
         return ApiResponse.error('Slider introuvable.');
       }
-
+      await this.sliderModel.findByIdAndDelete(idSlider);
       // 2. Suppression du fichier physique
       if (slider.image) {
         const fullPath = path.join(process.cwd(), slider.image);
-
         if (fs.existsSync(fullPath)) {
           try {
             fs.unlinkSync(fullPath);
           } catch (fileError) {
-            return ApiResponse.error(
-              'Une erreur est survenue lors de la suppression de l’image', fileError
-            );
+            console.error('Erreur fichier:', fileError);
           }
         }
       }
-
-      // 3. Suppression en base de données
-      const result = await this.sliderModel.deleteOne({ slug: slug });
       return ApiResponse.success('Slider supprimé avec succès.');
     }
     catch (error) {
